requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
confluent-kafka==2.4.0
idna==3.7
PyYAML==6.0.1
requests==2.32.3
urllib3==2.2.1
